+++ .gitignore
@@ -1,2 +1,4 @@
 *.pyc
 build/
+*.bin
+*.elf
